EUMM: avoid stderr noise if Win32 chcp not found
This is a point patch for RT #123998.

ExtUtils::MakeMaker::Locale on win32 uses various strategies to determine
the code page. One of these is to call out to the chcp command. If this
isn't found, a spurious error appears on stdout. As well as being visually
annoying, it causes t/porting/utils.t to fail:

    # Failed test 70 - utils/perlivp compiles at t/porting/utils.t line 85
    #      got "\'chcp\' is not recognized as an internal or external command,\noper
    able program or batch file.\nutils/perlivp syntax OK\n"
    # expected "utils/perlivp syntax OK\n"

This is a minimal fix that just redirects stderr to NUL.